+++ packages/react-sdk/src/client/SearchcraftPopoverForm.tsx
@@ -36,7 +36,7 @@ interface SearchcraftPopoverFormProps
  * ```
  */
 const SearchcraftPopoverForm: FC<SearchcraftPopoverFormProps> = (props) => (
-  <Component {...props} />
+  <Component {...props} sdkVariant='react' />
 );
 
 export { SearchcraftPopoverForm };